refactor: split internal crates
This is an initial refactor that splits a policy-fetcher crate out of
the policy-server.

Further refactors will follow reusing this logic.
