+++ crates/policy-fetcher/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "policy-fetcher"
+version = "0.1.0"
+authors = [
+  "Flavio Castelli <fcastelli@suse.com>",
+  "Rafael Fernández López <rfernandezlopez@suse.com>"
+]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+async-std = "1.9.0"
+async-stream = "0.3.0"
+async-trait = "0.1.42"
+base64 = "0.13.0"
+hyper-tls = "0.5.0"
+hyper = { version = "0.14", features = ["full"] }
+native-tls = "^0.2.1"
+oci-distribution = "0.5"
+serde = { version = "1.0", features = ["derive"] }
+serde_yaml = "0.8.15"
+tokio-compat-02 = "0.2.0"
+url = "2.2.0"
