--- src/registry/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod config;
